fix(GuildExpHistory): fixed possibly empty exp history
