+++ src/structures/Guild/GuildMember.js
@@ -1,3 +1,4 @@
+const {parseHistory} = require('../../utils/guildExp');
 /**
  * GuildMember class
  */
@@ -61,39 +62,5 @@ class GuildMember {
     return this.uuid;
   }
 }
-
-const dateRegExp = /(\d{4})-(\d{2})-(\d{2})/;
-/**
- * Parses exp history
- * @param {object} historyData History data from the API
- * @returns {Array<ExpHistory>} Array of ExpJistory
- */
-function parseHistory(historyData) {
-  return Object.entries(historyData).map((x, index) => ({
-    day: x[0],
-    date: parseDate(x[0].match(dateRegExp).slice(1).map((x) => parseInt(x, 10))) || undefined,
-    exp: x[1] || 0,
-    totalExp: Object.values(historyData).slice(0, index + 1).reduce((pV, cV) => pV + cV, 0)
-  }));
-}
-
-/**
- * Parses date
- * Because hypixel's oscillation precises that exp resets at 5 am UTC, the hour is set accordingly
- * @param {number[]} date Date from regexp
- * @returns {Date} Parsed Date
- */
-function parseDate(date) {
-  date[1] -= 1;
-  return new Date(Math.round(new Date(new Date().setUTCFullYear(...date)).setUTCHours(5, 0, 0) / 1000) * 1000);
-}
-
-/**
- * @typedef {object} ExpHistory
- * @property {string} day String Date ( unparsed )
- * @property {Date} date Parsed Date
- * @property {number} exp Experience of the day
- * @property {number} totalExp Experience earned from day 0 to this day
- */
 
 module.exports = GuildMember;
