+++ typings/index.d.ts
@@ -843,7 +843,7 @@ declare module 'hypixel-api-reborn' {
     weeklyExperience: number;
     mutedUntilTimestamp: number;
     mutedUntil: Date;
-    expHistory: { day: string, exp: number }[];
+    expHistory: { day: string, date: Date, exp: number, totalExp: number }[][];
   }
   class GuildRank {
     constructor(data: Record<string, unknown>);
